chore(core/modules): move config manager to config sub directory
