+++ apps/core/src/modules/config/config.manager.ts
